First version of mintTicket
Added a simple smart contract that is able to create NFT based on provided metadata, and previously to IFPS updated data. There is also a deploy script that deploys the smart contract and calls the mint function.

The next step will be to automate the process how creating the metadata etc.